AdminLTE view 적용

+++ src/main/java/com/kosta/cart/model/CartDAOImpl.java
@@ -20,12 +20,12 @@ public class CartDAOImpl implements CartDAO {
 	
 	@Override
 	public int checkInsertEbook(CartVO vo) {
-		return (int)sqlSession.selectOne("cart.checkInsertEbook", vo);
+		return sqlSession.selectOne("cart.checkInsertEbook", vo);
 	}
 	
 	@Override
 	public int checkInsertEbookSalelist(CartVO vo) {
-		return (int)sqlSession.selectOne("cart.checkInsertEbookSalelist", vo);
+		return sqlSession.selectOne("cart.checkInsertEbookSalelist", vo);
 	}
 
 	@Override
